This is a large-scale revision to the LCA code. I originally committed the code in the summer of 2019, but Jason P asked for significant changes including the ability to delete nodes. Also, fresh eyes in May 2020 led me to make some conceptual changes to the design. All of these changes are too hard to document in a single commit message so instead the code should be treated as new.
In approaching the code, start with the example.py file. As of
this writing, it is done for the main execution of the graph algorithm,
including gathering the weights of edges after they are pulled from
the databased, and establishment of weights from the output of a
verification algorithm such as VAMP and from the returned decisions made
by a human. The example needs more about running the ranking algorithm
prior to forming the weights and needs more about the analysis of the
decision results made by LCA.